% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dapmap.consumption}
\alias{dapmap.consumption}
\title{DAP/MAP consumption by region}
\format{
A data frame with 12 rows and 10 variables
\itemize{
  \item Year: year of operations
  \item Africa: dap/map consumption in countries
AGO, BDI, BEN, BFA, CAF, CIV, CMR, COD, COG, DJI, DZA, EGY, ERI, ETH, GAB, GHA, GIN, GMB, GNQ, KEN, LBR, LBY, MAR, MDG, MLI, MOZ, MRT, MUS, MWI, NAM, NER, NGA, RWA, SDN, SEN, SLE, SOM, SWZ, TCD, TGO, TUN, TZA, UGA, ZAF, ZMB, ZWE
  \item East Asia: dap/map consumption in countries
CHN, BRN, IDN, JPN, KHM, KOR, LAO, MMR, MNG, MYS, PHL, PRK, SGP, THA, TLS, VNM
  \item Eastern Europe and Central Asia: dap/map consumption in countries
ARM, AZE, BLR, EST, GEO, KAZ, KGZ, LTU, LVA, MDA, RUS, TJK, TKM, UKR, UZB
  \item Latin America: dap/map consumption in countries
ARG, ATG, BHS, BLZ, BMU, BOL, BRA, BRB, CHL, COL, CRI, CUB, CYM, DMA, DOM, ECU, GRD, GTM, GUY, HND, HTI, JAM, KNA, LCA, MEX, NIC, PAN, PER, PRY, SLV, SUR, TCA, TTO, URY, VCT, VEN
  \item North America: dap/map consumption in countries
CAN, USA
  \item Oceania: dap/map consumption in countries
AUS, FJI, NCL, NRU, NZL, PLW, PNG, PYF, SLB
  \item South Asia: dap/map consumption in countries
BGD, IND, LKA, NPL, PAK
  \item Western and Central Europe: dap/map consumption in countries
ALB, AUT, BEL, BGR, BIH, CHE, CZE, DEU, DNK, ESP, FIN, FRA, GBR, GRC, HRV, HUN, IRL, ISL, ITA, LUX, MKD, MNE, NLD, NOR, POL, PRT, ROU, SRB, SVK, SVN, SWE
  \item Western Asia: dap/map consumption in countries
AFG, ARE, BHR, CYP, IRN, IRQ, ISR, JOR, KWT, LBN, OMN, PSE, QAT, SAU, SYR, TUR, YEM
}
}
\usage{
dapmap.consumption
}
\description{
This compilation provides records on the DAP/MAP fertilizer consumption estimated from the IFA Supply Database.
}
\details{
The following definitions apply: DAP - diammonium phosphate, MAP - monoammonium phosphate.

This compilation covers the period from 2007 to 2018.

Countries are reported by ISO 3166 country code.

Values are exported as aggregated DAP and MAP apparent consumption by region in the IFA Supply Database.

Metric tonnes of P2O5 per annum are used for volumes throughout this compilation.
Original data values are recorded in metric tonnes of product.

The following conversion factors of product to P2O5 volumes were applied:
\itemize{
  \item DAP   0.46
  \item MAP   0.52
}
}
\references{
\itemize{
	 \item IFA Supply Database, downloaded from https://www.ifastat.org/supply/Phosphate\%20Products/Processed\%20Phosphates in Jun 2020
}
}
\keyword{datasets}
